requirements.txt
fastapi~=0.104.1
i18nice~=0.12
uvicorn~=0.15.0
